# Optimize `split_states`