--- a/lib/lrama/state.rb
+++ b/lib/lrama/state.rb
@@ -308,11 +308,7 @@ module Lrama
             [action, action.rule.empty_rule? ? lhs_contributions(action.rule.lhs, token) : kernels.map {|k| [k, k.end_of_rule?] }.to_h]
           end
         }.to_h
-        if (annotation = @annotation_list.find {|a| a.state == self && a.token == token && a.actions == actions })
-          annotation.merge_matrix(contribution_matrix)
-        else
-          @annotation_list << InadequacyAnnotation.new(self, token, actions, contribution_matrix)
-        end
+        @annotation_list << InadequacyAnnotation.new(self, token, actions, contribution_matrix)
       }
     end
 
@@ -321,8 +317,6 @@ module Lrama
     # @rbs (State predecessor) -> void
     def annotate_predecessor(predecessor)
       propagating_list = annotation_list.map {|annotation|
-        next nil if annotation.only_always_or_never?
-
         contribution_matrix = annotation.contribution_matrix.map {|action, contributions|
           if contributions.nil?
             [action, nil]
@@ -339,6 +333,7 @@ module Lrama
             [action, cs]
           end
         }.to_h
+        next nil if contribution_matrix.all? {|_, contributions| contributions.nil? || contributions.all? {|_, contributed| !contributed } }
 
         InadequacyAnnotation.new(annotation.state, annotation.token, annotation.actions, contribution_matrix)
       }.compact