- Remove unnecessary re-computation
    - `compute_follow_kernel_items`
    - `compute_always_follows`
    - `compute_goto_follows`
- Check if the contribution matrix is potential before propagation

```sh
$ time bundle exec lrama --trace=time -D lr.type=ielr parse.y
parse    0.43902 s
parse    0.00316 s
compute_lr0_states    0.40920 s
compute_direct_read_sets    0.07128 s
compute_reads_relation    0.01224 s
compute_read_sets    0.02581 s
compute_includes_relation    0.11171 s
compute_lookback_relation    0.23253 s
compute_follow_sets    0.08286 s
compute_look_ahead_sets    0.86599 s
compute_conflicts    0.00491 s
compute_default_reduction    0.00237 s
compute_predecessors    0.11527 s
compute_follow_kernel_items    0.47898 s
compute_always_follows    0.41872 s
compute_goto_follows    1.12686 s
compute_inadequacy_annotations   46.91071 s
split_states    2.11387 s
compute_direct_read_sets    0.02702 s
compute_reads_relation    0.00805 s
compute_read_sets    0.03454 s
compute_includes_relation    0.12045 s
compute_lookback_relation    0.23710 s
compute_follow_sets    0.15189 s
compute_look_ahead_sets    1.62029 s
compute_conflicts    0.00534 s
compute_default_reduction    0.00251 s
compute_yydefact    0.06543 s
compute_yydefgoto    0.00893 s
sort_actions    0.00624 s
compute_packed_table    0.33528 s
render    0.07173 s
bundle exec lrama --trace=time -D lr.type=ielr parse.y  56.52s user 0.23s system 99% cpu 57.030 total
```