@yui-knk 2025-03-12
```
split_states    2.11387 s
```

Great work!